requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
mysql-connector-python==8.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pyTelegramBotAPI==4.17.0
pytz==2024.1
requests==2.31.0
typing_extensions==4.11.0
urllib3==2.2.1
yarl==1.9.4
